commented

+++ src/AppliBundle/Controller/ProjetController.php
@@ -26,7 +26,10 @@ class ProjetController extends Controller
      */
     public function indexAction(User $user)
     {
+        // Controle de l'utilisateur
         if ($user === $this->getUser()){
+
+            //Récupération des données du projet et données liées
             $em = $this->getDoctrine()->getManager();
 
             $projets = $em->getRepository('AppliBundle:Projet')->findByUtilisateur($user);
@@ -40,13 +43,15 @@ class ProjetController extends Controller
 
             ));
         }
+
+        // Si mauvais Utilisateur on renvoie vers page Unauthorized
         else{
             return $this->render('AppliBundle:Default:unauthorized.html.twig');
         }
     }
 
     /**
-     * Creates a new projet entity.
+     * Creates a new projet entity with user ID
      *
      * @Route("/{id}/new", name="projet_new")
      *
@@ -54,57 +59,68 @@ class ProjetController extends Controller
      */
     public function newAction(Request $request, User $user)
     {
+        // Controle de l'utilisateur
         if ($user === $this->getUser()){
+
+            // Création du nouveau projet
             $projet = new Projet();
-        $projet->setUtilisateur($user);
-        $form = $this->createForm('AppliBundle\Form\ProjetType', $projet);
-        $form->handleRequest($request);
 
+            // Attribution de l'utilisateur au projet
+            $projet->setUtilisateur($user);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->persist($projet);
-            $em->flush($projet);
+            // Formulaire de création du projet
+            $form = $this->createForm('AppliBundle\Form\ProjetType', $projet);
+            $form->handleRequest($request);
 
-            return $this->redirectToRoute('script_index', array(
-                'id' => $user->getId(),
-                'projet' => $projet->getId(),
+            // Si Formulaire valide, on persiste dans la BD
+            if ($form->isSubmitted() && $form->isValid()) {
+                $em = $this->getDoctrine()->getManager();
+                $em->persist($projet);
+                $em->flush($projet);
+
+                // et on renvoie vers la création du script
+                return $this->redirectToRoute('script_index', array(
+                    'id' => $user->getId(),
+                    'projet' => $projet->getId(),
+                ));
+            }
+
+            // Si formulaire invalide, on recommence
+            return $this->render('projet/new.html.twig', array(
+                'user' => $user,
+                'projet' => $projet,
+                'form' => $form->createView(),
             ));
-        }
+            }
 
-        return $this->render('projet/new.html.twig', array(
-            'user' => $user,
-            'projet' => $projet,
-            'form' => $form->createView(),
-        ));
-        }
+        // Si mauvais Utilisateur on renvoie vers page Unauthorized
         else{
             return $this->render('AppliBundle:Default:unauthorized.html.twig');
         }
     }
-
-    /**
-     * Finds and displays a projet entity.
-     *
-     * @Route("/{id}/{projet}", name="projet_show")
-     * @Method("GET")
-     */
-    public function showAction(User $user, Projet $projet)
-    {
-        if ($user === $this->getUser()){
-
-        $deleteForm = $this->createDeleteForm($projet, $user);
-
-        return $this->render('projet/show.html.twig', array(
-            'user' => $user,
-            'projet' => $projet,
-            'delete_form' => $deleteForm->createView(),
-        ));
-        }
-        else{
-            return $this->render('AppliBundle:Default:unauthorized.html.twig');
-        }
-    }
+//
+//    /**
+//     * Finds and displays a projet entity.
+//     *
+//     * @Route("/{id}/{projet}", name="projet_show")
+//     * @Method("GET")
+//     */
+//    public function showAction(User $user, Projet $projet)
+//    {
+//        if ($user === $this->getUser()){
+//
+//        $deleteForm = $this->createDeleteForm($projet, $user);
+//
+//        return $this->render('projet/show.html.twig', array(
+//            'user' => $user,
+//            'projet' => $projet,
+//            'delete_form' => $deleteForm->createView(),
+//        ));
+//        }
+//        else{
+//            return $this->render('AppliBundle:Default:unauthorized.html.twig');
+//        }
+//    }
 
     /**
      * Displays a form to edit an existing projet entity.
@@ -114,24 +130,32 @@ class ProjetController extends Controller
      */
     public function editAction(User $user, Request $request, Projet $projet)
     {
+
+        // Controle de l'utilisateur
         if ($user === $this->getUser()){
-        $deleteForm = $this->createDeleteForm($projet, $user);
-        $editForm = $this->createForm('AppliBundle\Form\ProjetType', $projet);
-        $editForm->handleRequest($request);
 
-        if ($editForm->isSubmitted() && $editForm->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
 
-            return $this->redirectToRoute('projet_index', array('id' => $user->getId()));
+//        $deleteForm = $this->createDeleteForm($projet, $user);
+            // création du formulaire avec données remplies
+            $editForm = $this->createForm('AppliBundle\Form\ProjetType', $projet);
+            $editForm->handleRequest($request);
+
+            // si formulaire valide on modifie les données dans la BDD
+            if ($editForm->isSubmitted() && $editForm->isValid()) {
+                $this->getDoctrine()->getManager()->flush();
+
+                // et on redirige vers la liste des projets
+                return $this->redirectToRoute('projet_index', array('id' => $user->getId()));
         }
-
-        return $this->render('projet/edit.html.twig', array(
-            'user' => $user,
-            'projet' => $projet,
-            'edit_form' => $editForm->createView(),
-            'delete_form' => $deleteForm->createView(),
+            // sinon formulaire non valide on reessaie
+            return $this->render('projet/edit.html.twig', array(
+                'user' => $user,
+                'projet' => $projet,
+                'edit_form' => $editForm->createView(),
+//                'delete_form' => $deleteForm->createView(),
         ));
         }
+        // Si mauvais Utilisateur on renvoie vers page Unauthorized
         else{
             return $this->render('AppliBundle:Default:unauthorized.html.twig');
         }
@@ -145,17 +169,22 @@ class ProjetController extends Controller
      */
     public function deleteAction(User $user, Request $request, Projet $projet)
     {
-//        $form = $this->createDeleteForm($projet, $user);
-//        $form->handleRequest($request);
-//
-//        if ($form->isSubmitted() && $form->isValid()) {
+        // Controle de l'utilisateur
+        if ($user === $this->getUser()){
+
+            // Suppression dans la BDD du projet séléctionné
             $em = $this->getDoctrine()->getManager();
             $em->remove($projet);
             $em->flush($projet);
-//        }
-
-        return $this->redirectToRoute('projet_index', array('id' => $user->getId()));
+            // et on redirige vers la liste des projets
+            return $this->redirectToRoute('projet_index', array('id' => $user->getId()));
+        }
+        // Si mauvais Utilisateur on renvoie vers page Unauthorized
+        else{
+            return $this->render('AppliBundle:Default:unauthorized.html.twig');
+        }
     }
+
 
     /**
      * Creates a form to delete a projet entity.
